# perf: Improve `@babel/types` builders