--- a/packages/babel-types/src/validators/validate.ts
+++ b/packages/babel-types/src/validators/validate.ts
@@ -8,7 +8,7 @@ import type * as t from "../index.ts";
 export default function validate(
   node: t.Node | undefined | null,
   key: string,
-  val: any,
+  val: unknown,
 ): void {
   if (!node) return;
 
@@ -20,10 +20,29 @@ export default function validate(
   validateChild(node, key, val);
 }
 
+export function validateInternal(
+  field: FieldOptions,
+  node: t.Node | undefined | null,
+  key: string,
+  val: unknown,
+  maybeNode?: boolean,
+): void {
+  if (!field?.validate) return;
+  if (field.optional && val == null) return;
+
+  field.validate(node, key, val);
+
+  if (maybeNode) {
+    const type = node?.type;
+    if (type == null) return;
+    NODE_PARENT_VALIDATIONS[type]?.(node, key, val);
+  }
+}
+
 export function validateField(
   node: t.Node | undefined | null,
   key: string,
-  val: any,
+  val: unknown,
   field: FieldOptions | undefined | null,
 ): void {
   if (!field?.validate) return;
@@ -35,10 +54,9 @@ export function validateField(
 export function validateChild(
   node: t.Node | undefined | null,
   key: string,
-  val?: t.Node | undefined | null,
+  val?: unknown,
 ) {
-  if (val == null) return;
-  const validate = NODE_PARENT_VALIDATIONS[val.type];
-  if (!validate) return;
-  validate(node, key, val);
+  const type = node?.type;
+  if (type == null) return;
+  NODE_PARENT_VALIDATIONS[type]?.(node, key, val);
 }