--- a/scripts/utils/formatCode.js
+++ b/scripts/utils/formatCode.js
@@ -3,7 +3,11 @@ import * as prettier from "prettier";
 export default async function formatCode(code, filename) {
   const prettierConfig = await prettier.resolveConfig(filename);
   prettierConfig.filepath = filename;
-  prettierConfig.parser = filename.endsWith(".ts") ? "babel-ts" : "babel";
+  prettierConfig.parser = filename.endsWith(".json")
+    ? "json"
+    : filename.endsWith(".ts")
+      ? "babel-ts"
+      : "babel";
 
   return prettier.format(code, prettierConfig);
 }