<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
Some discussion is needed.
Because it's a breaking change.
Ref: https://github.com/babel/babel/issues/13868

This PR makes `fields` and `builderKeys` only lookup once on initialization.
```
PS F:\babel\benchmark\babel-types\builders> node .\memberExpression.mjs
baseline memberExpression builder: 804_478 ops/sec ±0.52% (0.001ms)
current memberExpression builder: 7_391_582 ops/sec ±0.26% (0ms)
PS F:\babel\benchmark\babel-types\builders> node .\stringLiteral.mjs
baseline stringLiteral builder: 32_462_576 ops/sec ±0.42% (0ms)
current stringLiteral builder: 81_001_155 ops/sec ±0.6% (0ms)
```


<a href="https://gitpod.io/#https://github.com/babel/babel/pull/15843"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>