@nicolo-ribaudo 2024-09-04
Lets do this without the suggestion I had in https://github.com/babel/babel/pull/15843#pullrequestreview-1565752684 :)