@liuxingbaoyu 2024-09-05
> Lets do this without the suggestion I had in [#15843 (review)](https://github.com/babel/babel/pull/15843#pullrequestreview-1565752684) :)

We seem to have done that in a subsequent commit. :)