@liuxingbaoyu 2024-09-05
Since `BABEL_8_BREAKING` is dynamic, this PR is a bit harder to merge.
I will open a PR to complete #13868, which will allow us to merge this PR in the Babel 8 patch release.