@liuxingbaoyu 2024-09-17
I'll try to land it in Babel 7, since #13868 is no longer blocking.
At the same time I realize that this will be annoying as long as we use `process.env.BABEL_8_BREAKING` or even `process.env.BABEL_9_BREAKING`.
Hopefully we can fix it once and for all.