@liuxingbaoyu 2024-09-18
I think the CI error is a build error in Babel 8 where we accidentally bundled `@babel/types` everywhere. 🤦‍♂️
![image](https://github.com/user-attachments/assets/183aaa9f-6ff9-4cc3-8da2-9865be2aa0db)